use TypeNames defined in the card component.
