--- src/fixtures/connector-type-data.ts
@@ -1,9 +1,0 @@
-interface TypeModel {
-    id: number;
-    name: string;
-}
-export const TYPE_NAMES_DATA: TypeModel[] = [
-    { id: 1, name: 'Foundation' },
-    { id: 2, name: 'CRM' },
-    { id: 3, name: 'Talent' },
-];
